+++ NEWS.md
@@ -23,4 +23,6 @@
 
 * Add informative error if NaN values are generated using sim_distribution
 * Allow for more flexible simulation of sets by adding an argument to sim_sets called subset_cells and a custom_sets argument to sim_survey
+* Add number of fish available to the survey (I = N * q) to the sp_N object when running sim_survey
+